Add ofInterest rule as a separate module
