+++ compiler/hie-core/src/Development/IDE/Core/OfInterest.hs
@@ -1,0 +1,81 @@
+-- Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+-- SPDX-License-Identifier: Apache-2.0
+
+
+{-# LANGUAGE TypeFamilies               #-}
+{-# LANGUAGE OverloadedStrings          #-}
+{-# LANGUAGE FlexibleInstances #-}
+
+-- | A Shake implementation of the compiler service, built
+--   using the "Shaker" abstraction layer for in-memory use.
+--
+module Development.IDE.Core.OfInterest(
+    ofInterestRules,
+    getFilesOfInterest, setFilesOfInterest, modifyFilesOfInterest,
+    ) where
+
+import           Control.Concurrent.Extra
+import           Control.Monad.Except
+import Data.Hashable
+import Control.DeepSeq
+import GHC.Generics
+import Data.Typeable
+import qualified Data.ByteString.UTF8 as BS
+import Control.Exception
+import Development.IDE.Types.Location
+import Development.IDE.Types.Logger
+import           Data.Set                                 (Set)
+import qualified Data.Set                                 as Set
+import qualified Data.Text as T
+import Data.Tuple.Extra
+import Development.IDE.Types.Location (NormalizedFilePath)
+import           Development.Shake                        hiding (Diagnostic, Env, newCache)
+
+import           Development.IDE.Core.Shake
+
+
+
+newtype OfInterestVar = OfInterestVar (Var (Set NormalizedFilePath))
+instance IsIdeGlobal OfInterestVar
+
+
+type instance RuleResult GetFilesOfInterest = Set NormalizedFilePath
+
+
+data GetFilesOfInterest = GetFilesOfInterest
+    deriving (Eq, Show, Typeable, Generic)
+instance Hashable GetFilesOfInterest
+instance NFData   GetFilesOfInterest
+
+
+ofInterestRules :: Rules ()
+ofInterestRules = do
+    defineEarlyCutoff $ \GetFilesOfInterest _file -> assert (null $ fromNormalizedFilePath _file) $ do
+        alwaysRerun
+        filesOfInterest <- getFilesOfInterestUntracked
+        pure (Just $ BS.fromString $ show filesOfInterest, ([], Just filesOfInterest))
+
+
+getFilesOfInterest :: Action (Set NormalizedFilePath)
+getFilesOfInterest = use_ GetFilesOfInterest ""
+
+
+
+------------------------------------------------------------
+-- Exposed API
+
+-- | Set the files-of-interest which will be built and kept-up-to-date.
+setFilesOfInterest :: IdeState -> Set NormalizedFilePath -> IO ()
+setFilesOfInterest state files = modifyFilesOfInterest state (const files)
+
+getFilesOfInterestUntracked :: Action (Set NormalizedFilePath)
+getFilesOfInterestUntracked = do
+    OfInterestVar var <- getIdeGlobalAction
+    liftIO $ readVar var
+
+modifyFilesOfInterest :: IdeState -> (Set NormalizedFilePath -> Set NormalizedFilePath) -> IO ()
+modifyFilesOfInterest state f = do
+    OfInterestVar var <- getIdeGlobalState state
+    files <- modifyVar var $ pure . dupe . f
+    logDebug (ideLogger state) $ "Set files of interest to: " <> T.pack (show $ Set.toList files)
+    void $ shakeRun state [] (const $ pure ())
